Fixed serialization bug causing improper read and write, sparseSet's sieralize and deserialize were reading/writing as uint32_t vs size_t causing the pointer to end up in random locations

+++ SparseSet.h
@@ -132,7 +132,7 @@ public:
 	{
 		if (inFile.is_open())
 		{
-			uint32_t numEntities;
+			size_t numEntities;
 			inFile.read(reinterpret_cast<char*>(&numEntities), sizeof(numEntities));
 
 			std::vector<uint32_t> dense;
